+++ src/burst/ICounter.java
@@ -4,7 +4,7 @@ import annotation.Lock;
 
 public interface ICounter {
     @Lock(type="r")
-    public int getCounter();
+    public long getCounter();
 
     @Lock(type="w")
     public void increaseCounter();
